Move DeleteBlogModal to a seperate file
