+++ src/components/BlogContextMenu.js
@@ -1,73 +1,8 @@
 import React from 'react';
 
-import { Dropdown, DropdownToggle, DropdownMenu, DropdownItem, Collapse, Modal,
-  ModalBody, ModalFooter, ModalHeader, Card, CardHeader, Row, Col, FormGroup,
-  Label, Input, CardBody, Button, } from 'reactstrap';
+import { Dropdown, DropdownToggle, DropdownMenu, DropdownItem, } from 'reactstrap';
 
-class BlogDeleteModal extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isOpen: Boolean(props.isOpen),
-    }
-    
-    this.close = this.close.bind(this);
-    this.confirm = this.confirm.bind(this);
-    this.keyPress = this.keyPress.bind(this);
-
-    if (this.state.isOpen) {
-      document.addEventListener("keyup", this.keyPress);
-    }
-  }
-
-  close() {
-    document.removeEventListener("keyup", this.keyPress);
-    this.setState({
-      isOpen: false,
-    });
-  }
-
-  keyPress(event) {
-    if (event.keyCode === 13) {
-      this.confirm();
-    }
-  }
-
-  confirm() {
-    csrfFetch(`/cube/blog/remove/${this.props.postID}`, {
-      method: 'DELETE',
-      headers: {},
-    }).then((response) => {
-      if (!response.ok) {
-        console.log(response);
-      } else {
-        window.location.href = '';
-      }
-    });
-  }
-
-  render() {
-    const {isOpen} = this.state;
-    return (
-      <>
-        <Modal isOpen={isOpen} toggle={this.close}>
-          <ModalHeader toggle={this.close}>Confirm Delete</ModalHeader>
-          <ModalBody>
-            <p>Are you sure you wish to delete this post? This action cannot be undone.</p>
-          </ModalBody>
-          <ModalFooter>
-            <Button color="danger" onClick={this.confirm}>
-              Delete
-            </Button>{' '}
-            <Button color="secondary" onClick={this.close}>
-              Close
-            </Button>
-          </ModalFooter>
-        </Modal>
-      </>
-    )
-  }
-}
+import BlogDeleteModal from './BlogDeleteModal';
 
 class BlogContextMenu extends React.Component {
   constructor(props) {
@@ -77,7 +12,7 @@ class BlogContextMenu extends React.Component {
     this.state = {
       dropdownOpen: false,
       collapseOpen: false,
-      deleteModalIsOpen: false,
+      deleteModalOpen: false,
     };
   }
 
@@ -90,7 +25,7 @@ class BlogContextMenu extends React.Component {
 
   openDeleteModal(){
     this.setState({
-      deleteModalIsOpen: true,
+      deleteModalOpen: true,
     });
   }
 
@@ -122,12 +57,20 @@ class BlogContextMenu extends React.Component {
           <DropdownItem onClick={() => this.clickEdit(this.props.post)}>Edit</DropdownItem>
           <DropdownItem onClick={this.openDeleteModal}>
             Delete
-            <BlogDeleteModal postID={this.props.post._id} isOpen={this.state.deleteModalIsOpen}>
+            <BlogDeleteModal postID={this.props.post._id} isOpen={this.state.deleteModalOpen}>
             </BlogDeleteModal>
           </DropdownItem>
         </DropdownMenu>
       </Dropdown>
     );
+  }
+
+  componentDidUpdate() {
+    if (this.state.deleteModalOpen) {
+      this.setState({
+        deleteModalOpen: false,
+      })
+    }
   }
 }
